+++ src/lab_m1/lab2/lab2.h
@@ -16,7 +16,7 @@ namespace m1
         void Init() override;
 
      private:
-        void CreateMesh(const char *name, const std::vector<VertexFormat> &vertices, const std::vector<unsigned short> &indices);
+        void CreateMesh(const char *name, const std::vector<VertexFormat> &vertices, const std::vector<unsigned int> &indices);
 
         void FrameStart() override;
         void Update(float deltaTimeSeconds) override;
